requirements.txt
streamlit
ollama
watchdog
google-generativeai
mistralai
python-dotenv
cryptography
requests
openai
redis
pyyaml
gitpython
pytest
pytest-cov